sql: move gossipccl.DisableMerges to gossip, call during idx backfill
This moves the DisableMerges helper from gossipccl to gossip, and adds a call to it during index backfills.

Release note: none.
