+++ pkg/gossip/disable_merges_test.go
@@ -1,18 +1,23 @@
 // Copyright 2018 The Cockroach Authors.
 //
-// Licensed as a CockroachDB Enterprise file under the Cockroach Community
-// License (the "License"); you may not use this file except in compliance with
-// the License. You may obtain a copy of the License at
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-//     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
 
-package gossipccl
+package gossip
 
 import (
 	"context"
 	"testing"
 
-	"github.com/cockroachdb/cockroach/pkg/gossip"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/cockroach/pkg/util/metric"
 	"github.com/cockroachdb/cockroach/pkg/util/stop"
@@ -32,14 +37,14 @@ func TestDisableMerges(t *testing.T) {
 	}
 	for _, c := range testCases {
 		t.Run("", func(t *testing.T) {
-			g := gossip.NewTest(1, nil /* rpcContext */, nil, /* grpcServer */
+			g := NewTest(1, nil /* rpcContext */, nil, /* grpcServer */
 				stopper, metric.NewRegistry())
 			ctx, cancel := context.WithCancel(context.Background())
 			defer cancel()
 
-			DisableMerges(ctx, g, c.tableIDs)
+			g.DisableMerges(ctx, c.tableIDs)
 			for _, id := range c.tableIDs {
-				key := gossip.MakeTableDisableMergesKey(id)
+				key := MakeTableDisableMergesKey(id)
 				if _, err := g.GetInfo(key); err != nil {
 					t.Fatalf("expected to find %s, but got %v", key, err)
 				}
